requirements.txt
flask==2.3.3
python-dotenv==1.0.0
replicate==0.1.0
Pillow==9.4.0
requests==2.28.1
# pip freeze > requirements.txt
